requirements.txt
playwright
supabase
requests
python-dotenv
